config/db_config.py
import os
from pymongo import MongoClient
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# MongoDB connection
MONGO_URI = os.getenv("MONGO_URI")
DB_NAME = os.getenv("DB_NAME")

client = MongoClient(MONGO_URI)
db = client[DB_NAME]

job_posting = db["job_posting"]
job_parser = db["job_parser"]
seekers = db["seekers"]
recommendations = db["recommendations"]